Draft logic to rotate ellipsoids in `create_ellipsoid`
Make the `create_ellipsoid` function to rotate the ellipsoid 90 degrees
on the right angles when semiaxes lengths are provided in arbitrary
order. An ellipsoid defined by `a`, `b`, `c` with no initial rotation
angles passed to the function will be oriented so its semiaxes match
with x, y, z directions. Any rotation angles passed to the function will
be added to those 90 degrees ones.